requirements.txt
fastapi
uvicorn
python-docx
python-multipart
google-generativeai
